+++ sparse/utils.py
@@ -165,36 +165,6 @@ def isscalar(x):
     return not isinstance(x, SparseArray) and np.isscalar(x)
 
 
-class PositinalArgumentPartial(object):
-    def __init__(self, func, pos, posargs):
-        self.pos = list(pos)
-        self.posargs = list(posargs)
-        self.func = func
-
-        self.n = len(pos)
-
-        self.__doc__ = func.__doc__
-
-    def __call__(self, *args, **kwargs):
-        j = 0
-        totargs = []
-
-        for i in range(len(args) + self.n):
-            if j >= self.n or i != self.pos[j]:
-                totargs.append(args[i - j])
-            else:
-                totargs.append(self.posargs[j])
-                j += 1
-
-        return self.func(*totargs, **kwargs)
-
-    def __str__(self):
-        return str(self.func)
-
-    def __repr__(self):
-        return repr(self.func)
-
-
 def random_value_array(value, fraction):
     def replace_values(n):
         i = int(n * fraction)